Add E2E tests, fixtures, and helpers
Introduce a comprehensive end-to-end test suite and utilities for the napari-deeplabcut plugin. Added pytest conftest fixtures that auto-mark tests as e2e and monkeypatch Qt dialogs (QMessageBox, QInputDialog) plus a controllable overwrite-confirmation stub. New tests cover layer coloring and color-scheme UI behavior, overwrite/merge/save semantics (including filling NaNs vs. conflicting edits), routing and provenance (ensuring saves target the correct H5, machine/GT separation, promotion flow and sidecar creation), and small signature/NaN-equality helpers. Also add utils for creating minimal DLC-like projects and HDF test data. Removed the older single test file test_e2e.py in favor of these modular tests.
